final layout (for now)

+++ src/Interfaces/Frontend/MotorController.cpp
@@ -41,11 +41,11 @@ void MotorController::generateValues() {
     if (test_value >= 100.0f) {
         test_value = 0.0f;
     }
-    static int test_fault = 0;
+    static float test_fault = 0.0f;
     emit newTestFault(test_fault);
-    if (test_fault == 0) {
-        test_fault = 1;
+    if (test_fault == 0.0) {
+        test_fault = 1.0;
     } else {
-        test_fault = 0;
+        test_fault = 0.0;
     }
 }
